+++ Evry/Payonaccount/etc/config.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../Store/etc/config.xsd">
+    <default>
+        <payment>
+            <payonaccount>
+                <payment_action>capture</payment_action><!-- You can use another methor like capture  -->
+                <model>Evry\Payonaccount\Model\PaymentMethod</model>
+                <active>1</active>
+                <title>Pay On Account</title>
+                <order_status>pending</order_status><!-- set default order status-->
+            </payonaccount>
+        </payment>
+    </default>
+</config>
