+++ Evry/Payonaccount/etc/module.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noTestSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="Evry_Payonaccount" setup_version="2.0.0" active="true">
+    </module>
+</config>
