+++ Evry/Payonaccount/view/frontend/web/js/view/payment/method-renderer/payonaccount.js
@@ -1,0 +1,14 @@
+define(
+    [
+        'Magento_Checkout/js/view/payment/default'
+    ],
+    function (Component) {
+        'use strict';
+ 
+        return Component.extend({
+            defaults: {
+                template: 'Evry_Payonaccount/payment/payonaccount'
+            }
+        });
+    }
+);
